Rename: myPage 연결 표 위치옮김(페이지는 나중에 지울 것임)
