+++ src/App.jsx
@@ -7,7 +7,7 @@ import CompanyDetailPage from "./pages/companyDetailPage/CompanyDetailPage";
 import InvestmentStatusPage from "./pages/investmentStatusPage/InvestmentStatusPage";
 import ComparisonStatusPage from "./pages/comparisonStatusPage/ComparisonStatusPage";
 import MyPage from "./pages/myPage/MyPage";
-import MyResultPage from "./pages/myResultPage/MyResultPage";
+import MyResultPage from "./components/myPageTable/MyResultPage";
 
 function App() {
   return (
@@ -20,6 +20,7 @@ function App() {
             <Route path="/investment" element={<InvestmentStatusPage />} />
             <Route path="/comparison" element={<ComparisonStatusPage />} />
             <Route path="/my-company" element={<MyPage />} />
+            {/* ↓ 지울 것임 */}
             <Route path="/comparison-result" element={<MyResultPage />} />
           </Route>
           <Route path="*" element={<Notfound />} />
