+++ src/components/myPageTable/ResultTable.module.scss
